requirements.txt
scikit-optimize==0.7.4
pandas==1.1.0
statsmodels==0.12.0
dataclasses==0.7; python_version < '3.7'
ete3==3.1.1
numpy==1.18.0
pytest==5.2.0
anytree==2.8.0
typing==3.7.4.1
scikit_learn==0.22.1
matplotlib==3.0.2; python_version < '3.8'
matplotlib==3.3.1; python_version == '3.8'

networkx==2.4
xgboost==1.0.1
Pillow==7.1.2
imageio==2.8.0
pydot==1.4.1
seaborn==0.9.0; python_version < '3.7'
hyperopt==0.2.4
joblib==0.17.0
tensorflow==2.1.0; python_version <= '3.7'
tensorflow==2.3.1; python_version == '3.8'
